# marketplace/server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
  description: Comprehensive RESTful API for marketplace operations with real-time capabilities
  contact:
    name: Marketplace API Support
    email: support@marketplace.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.marketplace.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/recommendations:
    get:
      tags: [Recommendations]
      summary: Get personalized recommendations
      operationId: getRecommendations
      security: []
      parameters:
        - name: userId
          in: query
          description: User ID for personalization
          schema:
            type: string
        - name: itemId
          in: query
          description: Item ID for similar items
          schema:
            type: string
        - name: type
          in: query
          description: Recommendation type
          schema:
            type: string
            enum: [collaborative, content-based, hybrid]
            default: hybrid
        - name: limit
          in: query
          description: Number of recommendations
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Personalized recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      recommendations:
                        type: array
                        items:
                          $ref: '#/components/schemas/Recommendation'

  # Auctions Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Item:
      type: object
      properties:
        id:
          type: string
          example: item-123
        title:
          type: string
          example: Vintage Leather Jacket
        description:
          type: string
          example: Authentic vintage leather jacket in excellent condition
        price:
          type: number
          format: float
          example: 89.99
        category:
          type: string
          example: clothing
        tags:
          type: array
          items:
            type: string
          example: [vintage, leather, jacket]
        images:
          type: array
          items:
            type: string
            format: uri
        condition:
          type: string
          enum: [new, like-new, good, fair, poor]
        location:
          type: object
          properties:
            country:
              type: string
            state:
              type: string
            city:
              type: string
            zipCode:
              type: string
        isActive:
          type: boolean
        sellerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        views:
          type: integer
        favorites:
          type: integer
...
    Recommendation:
      type: object
      properties:
        itemId:
          type: string
        score:
          type: number
          format: float
        type:
          type: string
          enum: [collaborative, content-based, hybrid]
        item:
          $ref: '#/components/schemas/Item'
        reason:
          type: string
        sources:
          type: array
          items:
            type: string
...